inc: Mount /data and /cache as ext4
We're already sort-of doing this because of CONFIG_EXT4_USE_FOR_EXT23,
so now it's explicit.  The only difference is that ext4's fake-ext3
turns off delayed allocation to appease apps with bad expectations about
fsync.  Gingerbread should be fine with that though.

Note this doesn't actually convert the filesystems.  They won't get new
ext4 features without using tune2fs or reformatting entirely.  But for
the tweakers who'd like to do so, here's their ticket.

Change-Id: I101dcf009d9257c3a35115f58b27ba19cd38b152